refactored items to be part of budget component
