+++ src/app/budget/budget.component.html
@@ -32,4 +32,3 @@
       </table>
   </div>
 </div> <!-- Items -->
-<button (click)="log()"></button>
