+++ src/app/home/home.component.html
@@ -27,6 +27,6 @@
     </div>
   </div>
   <div *ngIf="viewMode === 1">
-    <app-budget [selectedTab]="selectedTab"></app-budget>
+    <app-budget [userName]="userName" [selectedTab]="selectedTab"></app-budget>
   </div>
 </div>
