+++ src/app/home/home.component.ts
@@ -38,21 +38,6 @@ export interface Items extends Array<{}> {
           style({ padding: '3px 20px 3px 15px', opacity: '0' }),
           animate(500)
       ])
-    ]),
-    trigger('itemEnter', [
-      transition(':enter', [
-          style({ 
-            opacity: '0',
-            transform: 'translate(-400px)'
-         }),
-          animate(800)
-      ])
-    ]),
-    trigger('hide', [
-      transition(':enter', [
-          style({ opacity: '0.5' }),
-          animate(600)
-      ])
     ])
   ]
 })
@@ -84,6 +69,10 @@ export class HomeComponent implements OnInit, AfterViewChecked {
       this.vmm.viewMode
       .subscribe(vm => {
         this.viewMode = vm;
+      });
+      this.vmm.total
+      .subscribe(total => {
+        this.total = total;
       });
      }
 
@@ -134,6 +123,10 @@ export class HomeComponent implements OnInit, AfterViewChecked {
     })
   }
 
+  selectTab(i) {
+    this.selectedTab = i;
+  }
+
   deleteTab(index): void {
     let i = this.accounts.findIndex(element => {
       return element.id === index
@@ -148,55 +141,8 @@ export class HomeComponent implements OnInit, AfterViewChecked {
       });
   }
 
-  selectTab(i) {
-    if (i === 'all') {
-      this.selectedTab = 'all';   
-      var container = [];
-      this.accountsSub = this.storage.getAllUserAccounts(this.userName) // Get array of Account observables
-        .subscribe(accounts => { 
-          this.storage.getAllItemsForThisUser(accounts) // Get items by searching each Account's id
-            .subscribe(values => {
-              container.push(values);
-              this.items = this.storage.filterForDisplay(container);
-              this.countUpTotal();
-            });
-        });
-    } else {
-      this.selectedTab = i;
-      if (this.itemsSub) this.itemsSub.unsubscribe();
-      this.itemsSub = this.storage.getItemsForView(this.selectedTab)
-        .subscribe((resp: any) => {
-          this.items = this.storage.filterForDisplay(resp);
-          this.countUpTotal();
-        })
-    }
-  }
-
-  createItem(boolean):void {
-    let dialogRef = this.dialog.open(DialogCreateItemComponent, {
-      width: '500px',
-      height: '500px',
-      panelClass:"createModalDialog",
-      data: {
-        accountId: this.selectedTab,
-        income: boolean }
-    });
-  }
-
-  deleteItem(item) {
-    this.storage.deleteItem(item.accountId, item.id);
-    if (this.selectedTab === 'all') this.selectTab('all');
-  }
-
   intoDelete() {
     this.deleteMode = !this.deleteMode;
-  }
-
-  countUpTotal() {
-    let sum = 0;
-    this.items.incomes.forEach(obj => sum += obj.amount);
-    this.items.expenses.forEach(obj => sum -= obj.amount);
-    this.total = sum;
   }
 
   checkOverflow() {
