+++ src/app/modals/snack-bar/snack-bar.component.css
@@ -1,6 +1,6 @@
 .popup-container {
-    font-family: Lato;
-    font-size: 17px;
+    font-family: 'Fira Sans';
+    font-size: 16px;
     text-align: center;
 }
 .snackbar-text {
